Ask answer question

See merge request IF3250-2018-K01-I-4/WebAlumniSTEI!13

+++ app/Http/Controllers/MembersController.php
@@ -8,12 +8,15 @@ use Illuminate\Support\Facades\Mail;
 use App\Mail\SendReverificationEmail;
 use App\Member;
 use \Auth;
+use App\Post;
+use App\Question;
+use App\Answer;
 
 class MembersController extends Controller
 {
     public function __construct()
     {
-        $this->middleware('admin', ['except' => ['show', 'edit', 'update']]);
+        //$this->middleware('admin', ['except' => ['show', 'edit', 'update']]);
     }
 
     /**
@@ -23,8 +26,20 @@ class MembersController extends Controller
      */
     public function index()
     {
-        $members = Member::orderBy('nim','asc')->paginate(20);
-        return view('members.list')->with('members', $members);
+        $isMember = Auth::guard('member')->user() != null;
+        $isAdmin = Auth::user() != null && Auth::user()->IsAdmin == 1;
+
+        if(!($isMember || $isAdmin))
+            return redirect('/');
+
+        if ($isAdmin) {
+            $members = Member::orderBy('nim','asc')->paginate(20);
+            return view('members.list')->with('members', $members);
+        } else {
+            $members = Member::orderBy('nim','asc')->get();
+            return view('showmember')->with('members', $members);
+        }
+        
     }
 
     /**
@@ -55,16 +70,38 @@ class MembersController extends Controller
      * @return \Illuminate\Http\Response
      */
     public function show($id)
-    {        
-        $user = Member::find($id);
-        $user['countQuestions'] = $user->questions->count();
-        $user['countAnswers'] = $user->answers->count();
-        if($user !== null){
-            return view('admin.profile')->with('user', $user);
+    {
+        $isMember = Auth::guard('member')->user() != null;
+        $isAdmin = Auth::user() != null && Auth::user()->IsAdmin == 1;
+
+        if(!($isMember || $isAdmin))
+            return redirect('/');
+        
+        if ($isAdmin) {
+            $user = Member::find($id);
+            $user['countQuestions'] = $user->questions->count();
+            $user['countAnswers'] = $user->answers->count();
+            if($user !== null){
+                return view('admin.profile')->with('user', $user);
+            } else {
+                return abort(404);
+            }
         } else {
-            return abort(404);
-        }
+            $posts = Post::where('user_id', $id)->get();
+            $questions = Question::where('member_id', $id)->get();
+            $answers = Answer::where('member_id', $id)->get();
+            
+            $user = Member::find($id);
+            if($user == null) {
+                return abort(404);
+            }
+
+            $userdata = [$posts, $questions, $answers, $user];
+
+            return view('admin.profile')->with('userdata', $userdata);   
+        }     
     }
+
 
     /**
      * Show the form for editing the specified resource.
@@ -83,7 +120,7 @@ class MembersController extends Controller
             return redirect('/');
         }
     }
-
+    
     /**
      * Update the specified resource in storage.
      *
@@ -150,7 +187,11 @@ class MembersController extends Controller
             Mail::to($user)->send(new SendReverificationEmail($token));
             return redirect('/members/' . $id)->with('success', 'Profile Updated. Confirmation code has been sent to new email.');
         }else{*/
+        if ($isAdmin) {
             return redirect('/admin/members/' . $id)->with('success', 'Profile Updated');
+        } else {
+            return redirect('/members/' . $id)->with('success', 'Profile Updated');
+        }
         //}
     }
 
@@ -162,6 +203,11 @@ class MembersController extends Controller
      */
     public function destroy($id)
     {
+        $isAdmin = Auth::user() != null && Auth::user()->IsAdmin == 1;
+
+        if(!$isAdmin)
+            return redirect('/');
+
         $user = Member::find($id);
         if($user !== null) {
             $user->delete();
